+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-05-26 22:37:43
-// Commit: fix(perf): fix memory leak by cancelling perf subscription
+// Updated at: 2026-05-27 09:40:49
+// Commit: test(perf): implement widget integration tests for perf
 
 class perfComponentState {
-  final String commitRef = '1480';
-  final String lastMessage = 'fix(perf): fix memory leak by cancelling perf subscription';
-  final DateTime timestamp = DateTime.parse('2026-05-26T22:37:43+05:00');
+  final String commitRef = '1481';
+  final String lastMessage = 'test(perf): implement widget integration tests for perf';
+  final DateTime timestamp = DateTime.parse('2026-05-27T09:40:49+05:00');
 }